app/config/settings.py
from pathlib import Path
import yaml

BASE_DIR = Path(__file__).resolve().parents[2]
REGISTRY_PATH = BASE_DIR / "app" / "config" / "source_registry.yaml"

def load_source_registry() -> dict:
    with open(REGISTRY_PATH, "r", encoding="utf-8") as f:
        return yaml.safe_load(f)
