requirements.txt
pydantic>=2.7
PyYAML>=6.0
pypdf>=4.2.0
python-docx>=1.1.0
streamlit>=1.36.0
pandas>=2.2.0
openai>=1.99.0
